requirements.txt
appnope==0.1.0
bleach==2.1.2
certifi==2018.1.18
chardet==3.0.4
click==6.7
decorator==4.2.1
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.0.2
django-debug-toolbar==1.9.1
django-extensions==2.0.0
django-test-without-migrations==0.6
django-versatileimagefield==1.9
entrypoints==0.2.3
getgist==0.1.2
gunicorn==19.7.1
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
python-decouple==3.1
pytz==2018.3
pyzmq==17.0.0
requests==2.18.4
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sqlformatter==1.3
sqlparse==0.2.4
static3==0.7.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
typing==3.6.4
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.4
